Cleaned up package and user factories
Created factories to better manage the stages of a package/user
(unapproved user, approved user, full user with package associations,
etc).

Setup spork/guard-spork for to run faster specs

Renamed Package#received to Package#received_on

Set default for User#approved to be false.

Deleted specs for PackagesHelper and PackagesController until
functionality needs to be added to the controllers to be tested.