Setup basic recipient interaction with packages
Recipients can login and view their packages and click a 'sign out'
button to verify they have received the package.  Added some helper
methods to Package and PackagesHelper to assist in building the view.

Also updated support files for logging in to work with webkit driver.
Instead of manually creating a user session, the process of visiting the
login page and signing in is done to ensure that component of the site
still works.