Add Delivery and Receipt models
In stepping back and reconsidering the existing approach taken, it makes
sense to extract some repetitive information in the packages table into
different, higher level models.

Deliveries serve to encompass an entire set of packages a worker will be
mailing out at a time. Receipts encompass a set of packages received for
a recipient within a delivery. Recipients then own delivery receipts
instead of individual packages, workers can search across deliveries and
much more.

This also tackles a problem found when designing the main notification
interface in terms of how to build the notification form and was the
primary motivation for the change.