Move functionality of UserSearch view into DeliveryForm view
When creating UserSearch I had planned on reusing it across the
application in various places where auto-completing users would be
helpful.  As I've started to work on these features (user signups), I
started to try to bend UserSearch to work in different scenarios which
was gross and hard to test.

The way I initially set it up only made sense in the context of
searching for recipients to created deliveries so I've moved the
functionality into the DeliveryForm view.