Restructure DateExtentions and Besko.extendDate()
Originally designed DateExtensions to look and work like a mixin for
date objects. Problems were that it was clunky to always extend dates
with it whenever I wanted that functionality.  I added
Besko.extendDate() to make that easier.

This refactor is the logical conclusion to make DateExtensions a method
on Besko that returns a new date object with all the extra extensions
built in. The extensions are isolated in a closure so the global
namespace is no longer being polluted with DateExtensions.

Besko.Date() also corrects for timezone offets when creating a date from
an ISO formatted string that does not include the time information.

Breaks functionality in delivery search currently.