Fixup and refactor notification.js
Render an anchor tag if a notification was provided by the server to
render on the page. Refactor notification.js to define most of it's
funcitonality in a closure, lowering it's public interface. Fixed to
re-show notifications container if previously closed.

Move 'error' and 'notice' methods into Besko.Support.

Tests were refactored since reloading fixtures causes the dom object
that the application javascript binds to to be replaced. Instead run all
expectations in one 'it' declaration.