Leave user unconfirmed until delivery is created
User#skip_confirmation! auto-confirms the user which is not the desired
behavior of creating the user without sending a confirmation email, but
confirming them if they are part of a persisted delivery.

Add User#skip_confirmation_email! that will set a flag to skip the
email to be sent by overriding
Devise::Models::Confirmable#send_on_create_confirmation_instructions