Switch `limit` option to `size` (#1)
Ldaptic does not attempt to pass down the `limit` option to the
adapters, but instead uses it to determine when to return from
processing results.

It does pass down the same options used in Ldaptic::Methods#search to
the adapter and connection so it is possible to pass the correct option
for limiting the number of results returned by the server.