Improve dev/prod parity around Shibboleth sign-in
Handle redirect inside a Rails controller instead of having it be
handled by Apache.

Only enable signing in via Shibboleth if the `SHIBBOLETH_LOGIN_HANDLER`
variable is set. Don't render the form to the user if the variable is
not set.

To simplify conditional styling, swap placement of Touchstone/Shibboleth
form and form to sign in via `devise`.