Add filtering by deliverer
Use `paraphrase` gem to simplify query logic.

Add constants in `DeliveryQuery` to map keywords to query param values.
This avoids hard-coding query param values throughout the codebase and
allows changing the values with minimal code churn or breaking
functionality.

Switch sorting feature test to use desk worker names to assert order
since the `appear_before` matcher finds 'UPS' and 'LaserShip' in the
`<select>` first. It isn't possible to limit the assertion to a part of
the page with `within` since the orderly gem uses `page.body` to make
the assertion[1].

[1] https://github.com/jmondo/orderly/blob/master/lib/orderly.rb#L8