Remove unneeded helper methods in `DeliveriesController`
Leverage `paraphrase` to filter and specify defaults for query params if
none are specified removing the need for `delivery_search_params` in
favor of using `@query` in the view to read param values.

Remove call to `Delivery` constant in `deliveries/new.html.haml`. Build
a delivery with receipts for each recipient in the query params in the
controller removing the need for `receipts_for_new_delivery`.